+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ### Added
 - Possibility to configure the default volume with the config.php key `music.default_volume`
   [#109](https://github.com/nc-music/music/issues/109)
+- Composer and BPM support [#123](https://github.com/nc-music/music/pull/123):
+  * Scan `composer` and `BPM` from the track metadata and store them in the database (rescan needed) @ineednousername
+  * Subsonic API: Include OpenSubsonic properties `bpm`, `displayComposer`, and `contributors[role='composer']` in all song responses @ineednousername
+  * Ampache API: Include property `composer` in all song reponses
+  * Advanced search: Searching tracks by composer name and BPM
+  * Smart playlists: Selecting tracks by composer
+  * Viewing composer details in the artist details pane
 
 ### Changed
 - Show some error details in the browser console when subscribing a podcast channel fails